# home/src/IO.Swagger/wwwroot/swagger-original.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "The API for the Home Starter project",
    "version" : "1.0.3",
    "title" : "home"
  },
  "host" : "virtserver.swaggerhub.com",
  "basePath" : "/motta/home/1.0.3",
  "schemes" : [ "https" ],
  "consumes" : [ "application/json" ],
  "produces" : [ "application/json" ],
  "paths" : {
...
    "/remotes" : {
      "get" : {
        "tags" : [ "Remote" ],
        "description" : "returns all installed remotes",
        "operationId" : "getRemotes",
        "parameters" : [ {
          "name" : "skip",
          "in" : "query",
          "description" : "number of records to skip",
          "required" : false,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "limit",
          "in" : "query",
          "description" : "max number of records to return",
          "required" : false,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "All the installed remotes",
            "schema" : {
              "type" : "array",
              "items" : {
                "type" : "string"
              }
            }
          }
        }
      }
    },
    "/remotes/{remote}" : {
      "get" : {
        "tags" : [ "Remote" ],
        "description" : "returns all ir codes from remote",
        "operationId" : "getRemoteCodes",
        "parameters" : [ {
          "name" : "remote",
          "in" : "path",
          "description" : "Lirc remote",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "All the codes",
            "schema" : {
              "type" : "array",
              "items" : {
                "type" : "string"
              }
            }
          }
        }
      }
    },
    "/remotes/{remote}/{code}" : {
      "get" : {
        "tags" : [ "Remote" ],
        "description" : "returns ir code from remote",
        "operationId" : "getRemoteCode",
        "parameters" : [ {
          "name" : "remote",
          "in" : "path",
          "description" : "Lirc remote",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "code",
          "in" : "path",
          "description" : "ir code",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "All the codes",
            "schema" : {
              "type" : "array",
              "items" : {
                "type" : "string"
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Remote" ],
        "description" : "flashes ir code simulating the remote control",
        "operationId" : "sendRemoteCode",
        "parameters" : [ {
          "name" : "remote",
          "in" : "path",
          "description" : "Lirc remote",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "code",
          "in" : "path",
          "description" : "ir code",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "response",
            "schema" : {
              "$ref" : "#/definitions/ApiResponse"
            }
          }
        },
        "x-swagger-router-controller" : "Remote"
      }
    },
    "/lighting/dimmers/{deviceId}/{value}" : {
      "post" : {
        "tags" : [ "Z-Wave" ],
        "operationId" : "setDimmer",
        "parameters" : [ {
          "name" : "deviceId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "value",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "maximum" : 100.0,
          "minimum" : 0.0,
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "response",
            "schema" : {
              "$ref" : "#/definitions/ApiResponse"
            }
          }
        },
        "x-swagger-router-controller" : "ZWave"
      }
    },
    "/lighting/dimmers/{deviceId}/{value}/timer/{timeunit}" : {
      "post" : {
        "tags" : [ "Z-Wave" ],
        "description" : "sets a dimmer to a specific value on a timer",
        "operationId" : "setDimmerTimer",
        "parameters" : [ {
          "name" : "deviceId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "value",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "timeunit",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "units",
          "in" : "query",
          "required" : false,
          "type" : "string",
          "default" : "milliseconds",
          "enum" : [ "seconds", "minutes", "milliseconds" ]
        } ],
        "responses" : {
          "200" : {
            "description" : "response",
            "schema" : {
              "$ref" : "#/definitions/ApiResponse"
            }
          }
        },
        "x-swagger-router-controller" : "ZWave"
      }
    },
    "/lighting/switches/{deviceId}" : {
      "get" : {
        "tags" : [ "Z-Wave" ],
        "operationId" : "getSwitchState",
        "parameters" : [ {
          "name" : "deviceId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "response",
            "schema" : {
              "$ref" : "#/definitions/DeviceState"
            }
          }
        },
        "x-swagger-router-controller" : "ZWave"
      }
    },
    "/lighting/switches/{deviceId}/{value}" : {
      "post" : {
        "tags" : [ "Z-Wave" ],
        "operationId" : "setSwitch",
        "parameters" : [ {
          "name" : "deviceId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "value",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "enum" : [ "true", "false" ]
        } ],
        "responses" : {
          "200" : {
            "description" : "response",
            "schema" : {
              "$ref" : "#/definitions/ApiResponse"
            }
          }
        },
        "x-swagger-router-controller" : "ZWave"
      }
    },
    "/lighting/switches/{deviceId}/{value}/timer/{minutes}" : {
      "post" : {
        "tags" : [ "Z-Wave" ],
        "description" : "sets a switch to a specific value on a timer",
        "operationId" : "setSwitchTimer",
        "parameters" : [ {
          "name" : "deviceId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "value",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "enum" : [ "true", "false" ]
        }, {
          "name" : "minutes",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "response",
            "schema" : {
              "$ref" : "#/definitions/ApiResponse"
            }
          }
        },
        "x-swagger-router-controller" : "ZWave"
      }
    },
...
  },
  "definitions" : {
...
    "ApiResponse" : {
      "type" : "object",
      "properties" : {
        "code" : {
          "type" : "integer",
          "format" : "int32"
        },
        "message" : {
          "type" : "string",
          "example" : "everything is ok"
        }
      },
      "example" : {
        "code" : 0,
        "message" : "everything is ok"
      }
    },
...
    "DeviceState" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "string"
        },
        "name" : {
          "type" : "string"
        },
        "lastUpdate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "level" : {
          "type" : "integer",
          "format" : "int32"
        }
      },
      "example" : {
        "level" : 0,
        "lastUpdate" : "2000-01-23T04:56:07.000+00:00",
        "name" : "name",
        "id" : "id"
      }
    },
...
  }
}